Creating Service interfaces to improve maintainability
